fech: create upload image component on uploadImage.vue
